use specific rustfmt nightly (#51)
rust-lang/rust#51699 has broken the nightly Rust toolchain. Our Travis configuration allows failures when building with that toolchain, but we use nightly rustfmt when building with all toolchains, so the bustage affects our stable and beta builds too.

To work around the problem, this branch uses a specific rustfmt nightly rather than the latest one. (We might want to keep doing this even after the bustage is fixed, to avoid such bustage in the future.)