reference dependencies by docs.rs URLs that work both locally and on docs.rs
I thought it would be useful to reference the docs that `cargo doc` creates for dependencies, since those docs apply to the specific versions of the dependencies that rkv uses.  But that breaks on docs.rs, which doesn't build docs for dependencies when it builds docs for a crate.

So this change refers to those dependencies by their docs.rs URLs, which will work both locally and on docs.rs.  It's slightly suboptimal that the references are to the latest versions of the dependency docs, but pointing to a specific version of the docs on docs.rs is highly likely to become out of date as rkv updates to newer versions of those dependencies.

This seems to be the best compromise.