update authors list to reflect significant contributions
ncloudioj and I have made significant contributions to the project over the last few months. We've both exceeded the commit count of the original author! So this adds us to the list of authors. (Unsure of convention, but there's clearly intent for the list to include multiple authors, and it seems reasonable for it to reflect the major contributors.)